+++ packages/wallet/src/initialization/instances/index.ts
@@ -1,1 +1,2 @@
+export { approvalController } from './approval-controller';
 export { keyringController } from './keyring-controller';